requirements.txt
numpy==1.13.3
torchsummary
easydict
matplotlib
pandas
split-folders
